Move AssemblyName to shared (#23737)
